backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
openai-whisper
torch>=2.4
pytest>=8.0,<9.0
